+++ OpponentsView.cpp
@@ -274,7 +274,7 @@ void COpponentsView::OnMiscCmd1()
 *******************************************************************************
 */
 
-CString COpponentsView::GetCellData(int nColumn, CRow& oRow, int nField)
+CString COpponentsView::GetCellData(size_t nColumn, CRow& oRow, size_t nField)
 {
 	// Compose full name if name column.
 	if (nColumn == SECRETARY)
